+++ patches/src/main/kotlin/app/revanced/patches/music/misc/extension/SharedExtensionPatch.kt
@@ -4,6 +4,5 @@ import app.revanced.patches.music.misc.extension.hooks.applicationInitHook
 import app.revanced.patches.shared.misc.extension.sharedExtensionPatch
 
 val sharedExtensionPatch = sharedExtensionPatch(
-    "music",
     applicationInitHook,
 )